refactor: changed package name. editor to command
