+++ api-server/src/main/kotlin/io/mustelidae/seaotter/domain/command/Option.kt
@@ -1,0 +1,3 @@
+package io.mustelidae.seaotter.domain.command
+
+interface Option
