+++ api-server/src/main/kotlin/io/mustelidae/seaotter/domain/command/RotateCommand.kt
@@ -1,4 +1,4 @@
-package io.mustelidae.seaotter.domain.editor
+package io.mustelidae.seaotter.domain.command
 
 import io.mustelidae.seaotter.domain.image.ImgscalrFlabbyImage
 import io.mustelidae.seaotter.domain.image.ThumbnailatorFlabbyImage
