+++ api-server/src/main/kotlin/io/mustelidae/seaotter/domain/command/RotateOption.kt
@@ -1,4 +1,4 @@
-package io.mustelidae.seaotter.domain.editor
+package io.mustelidae.seaotter.domain.command
 
 class RotateOption(
     val angle: Double?,
