--- api-server/src/main/kotlin/io/mustelidae/seaotter/domain/editor/Option.kt
@@ -1,3 +1,0 @@
-package io.mustelidae.seaotter.domain.editor
-
-interface Option
